+++ HappyTravel.Edo.UnitTests/Tests/Services/Agents/BookingRequestExecutorTests/ErrorTests.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Threading.Tasks;
+﻿using System.Threading.Tasks;
 using HappyTravel.Edo.Api.Infrastructure;
 using HappyTravel.Edo.Api.Services.Accommodations.Bookings.BookingExecution;
 using HappyTravel.Edo.Api.Services.Accommodations.Bookings.Management;
@@ -7,63 +6,55 @@ using HappyTravel.Edo.Api.Services.Accommodations.Bookings.ResponseProcessing;
 using HappyTravel.Edo.Api.Services.Analytics;
 using HappyTravel.Edo.Api.Services.Connectors;
 using HappyTravel.Edo.Data.Bookings;
+using HappyTravel.EdoContracts.Accommodations.Enums;
 using HappyTravel.SuppliersCatalog;
 using Microsoft.Extensions.Logging;
 using Moq;
 using Xunit;
-using BookingRequest = HappyTravel.EdoContracts.Accommodations.BookingRequest;
 
 namespace HappyTravel.Edo.UnitTests.Tests.Services.Agents.BookingRequestExecutorTests
 {
-    public class BookingRequestExecutorConnectorsTests
+    public class ErrorTests
     {
         [Theory]
-        [InlineData(Suppliers.Illusions)]
-        [InlineData(Suppliers.Etg)]
-        [InlineData(Suppliers.Columbus)]
-        public async Task Correct_supplier_should_be_passed_when_getting_supplier_connector(Suppliers supplier)
+        [InlineData(BookingFailureCodes.ConnectorValidationFailed, true, false)]
+        [InlineData(BookingFailureCodes.ValuationResultNotFound, true, false)]
+        [InlineData(BookingFailureCodes.PreBookingFailed, true, false)]
+        [InlineData(BookingFailureCodes.SupplierValidationFailed, true, false)]
+        [InlineData(BookingFailureCodes.SupplierRejected, true, false)]
+        [InlineData(BookingFailureCodes.RequestException, true, true)]
+        [InlineData(BookingFailureCodes.UnknownRequestError, true, true)]
+        [InlineData(BookingFailureCodes.Unknown, false, true)]
+        public async Task Booking_failure_code_should_lead_to_failure_or_success_depending_on_error_code(
+            BookingFailureCodes failureCode,
+            bool isErrorDefined,
+            bool isSuccess)
         {
             InitializeMocks();
             var service = CreateBookingRequestExecutor();
-            var booking = new Booking { Supplier = supplier };
+            var booking = new Booking();
             var request = Utility.CreateAccommodationBookingRequest();
+            var problemDetails = Utility.CreateProblemDetailsWithFailureCode(isErrorDefined, failureCode);
+            Utility.SetupConnectorBookFailure(_supplierConnectorMock, problemDetails);
 
-            await service.Execute(request, default, booking, default, default);
+            var result = await service.Execute(request, default, booking, default, default);
 
-            _supplierConnectorManagerMock.Verify(x => x.Get(supplier));
+            Assert.Equal(isSuccess, result.IsSuccess);
         }
 
 
-        [Theory]
-        [InlineData("1", "2", "aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa", true)]
-        [InlineData("3", "4", "aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaab", false)]
-        [InlineData("5", "6", "aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaac", true)]
-        public async Task Correct_request_should_be_passed_when_booking_on_connector(
-            string availabilityId,
-            string referenceCode,
-            string roomContractSetIdString,
-            bool rejectIfUnavailable)
+        [Fact]
+        public async Task Exception_should_lead_to_success()
         {
             InitializeMocks();
-            var roomContractSetId = new Guid(roomContractSetIdString);
             var service = CreateBookingRequestExecutor();
-            var booking = new Booking { ReferenceCode = referenceCode };
-            var request = Utility.CreateAccommodationBookingRequest(roomContractSetId, rejectIfUnavailable);
-            BookingRequest actualInnerRequest = default;
-            var setupResult = Utility.SetupConnectorBookSuccess(_supplierConnectorMock);
-            SaveConnectorBookPassedParameter();
+            var booking = new Booking();
+            var request = Utility.CreateAccommodationBookingRequest();
+            Utility.SetupConnectorBookThrowsException(_supplierConnectorMock);
 
-            await service.Execute(request, availabilityId, booking, default, default);
+            var result = await service.Execute(request, default, booking, default, default);
 
-            _supplierConnectorMock.Verify(x => x.Book(It.IsAny<BookingRequest>(), It.IsAny<string>()));
-            Assert.Equal(availabilityId, actualInnerRequest.AvailabilityId);
-            Assert.Equal(roomContractSetId, actualInnerRequest.RoomContractSetId);
-            Assert.Equal(rejectIfUnavailable, actualInnerRequest.RejectIfUnavailable);
-            Assert.Equal(referenceCode, actualInnerRequest.ReferenceCode);
-
-            
-            void SaveConnectorBookPassedParameter()
-                => setupResult.Callback<BookingRequest, string>((req, _) => actualInnerRequest = req);
+            Assert.True(result.IsSuccess);
         }
